+++ CHANGELOG.md
@@ -8,7 +8,9 @@
 
 #### Bug修复
 
-- (无)
+- 修正 Hibernate 实体字段列注解中错误的列名。
+  - com.dwarfeng.scg.impl.bean.entity.HibernateNodeVariable.stringValue。
+  - com.dwarfeng.scg.impl.bean.entity.HibernateCommonVariable.stringValue。
 
 #### 功能移除